Delete calendar cache for users that have the 'read_private_events' cap.
Calendar views are cached based on the logged-in user's capability.  This
interferes with the AJAX calendar view in groups for these users, namely
admins.

Todo: Transient calendar caching in EO needs to be overhauled or, at the
very least, the transient key needs to be more unique and less generic.