Improve the filtering of group events when viewing calendars.
The previous implementation queried for all events, and then filtered out
individual events not belonging to the group using the `intercept_calendar()`
filter on 'eventorganiser_fullcalendar_event'. This technique is inefficent,
especially as the number of group events increases.

The new technique is to detect the 'bp_group_id' param sent with the AJAX
request, and then add a corresponding 'bp_group' param to the event query, using
the new 'eventorganiser_fullcalendar_query' filter in EO. This mirrors the way
user-specific calendar requests are filtered.

See #2, #5, #8.