Groups: On a single event page, ensure that the "Connected Groups" block is shown for a recurring event.
In the event-meta-event-single.php template, when the single event is
recurring, a WP_Query loop is used to output the number of occurences.  At
the end of the recurrence loop, wp_reset_postdata() is used.

wp_reset_postdata() interferes with BuddyPress by reinstating the current
post back to page ID 0.  For the 'eventorganiser_additional_event_meta'
hook to work properly (our "Connected Groups" block), we do not want the
page reinstated back to page ID 0, we want to reference the single event
page.

In order to do this, we have to manipulate the $wp_query global at the
end of the recurrence post loop.  Yay for loops within loops within loops!

See #7.