WP Frontend Admin Screen: Fix convert_to_screen().
To fix issues with plugins adding meta boxes that are not part of our
post type, we shouldn't use get_current_screen() in our version of
convert_to_screen().

Instead, we use the original unique id passed by add_meta_box().  This is
to ensure that meta boxes are displayed based on the proper post type.

See #6.