Merge single event content display logic.
Previously, we were duplicating the single event content on user event and
group event pages.  This wasn't so great.

This commit splits up the single event content and action links into
separate functions and out of the BP display methods.

This tidies things up and opens the door to move to a template part.