Yet Another Fricking Attempt to get activity posting to work in a consistent way.
I've modified the frontend abstraction class to introduce an `after_save()`
method. This allows me to fire the same hooks during creation through the
frontend - 'eventorganiser_created_event' and 'eventorganiser_updated_event' -
as are fired during creation via the EO API functions. This makes it easier
to create events in a systematic way.

To make this work, I've introduced a horrific workaround that detects when
`eventorganiser_details_save()` is being called, and ensures that no activity
item is created at that time.

See #23.