Users: Run single event logic on 'bp_parse_query'.
Previously, the single event logic on a user's profile ran on
'bp_setup_globals'.  Since EO registers their post type on 'init', there is
the potential that EO's post type is not registered before fetching the
single event and doing checks on the logged-in user's cap.

This commit moves the single event look up to 'bp_parse_query', which
should ensure that EO's post type is registered before we pull the
single event.

See #13.