iCal: Allow private group members to view contents of feed.
Previously, if an event was published in a private group, only the creator
of the event could view the contents of the iCal feed.  This is due to how
WordPress limits private posts to the author of the post.

This commit opens access up to logged-in group members, which should allow
these users to view the iCal feed as intended.