iCal: Add links to download public/private iCalendars on a user's 'Events' page.
At the bottom of a user's event calendar, the user will be able to download
both public and private iCalendars.

A private iCalendar includes both public and private events.  This can be
shared with friends.  If a user wants to reset a private iCalendar URL,
this can be done on a user's 'Events > Manage' page.

See #22.