Don't register CSS and JS assets until they're needed.
Registering Select2 early (and globally) led to problems when other plugins
register their own, out-of-date versions of Select2, since it creates a race
condition where the first plugin loaded gets preferential treatment for its
preferred version of Select2. To mitigate the extent of the problem, we're
switching to a technique where Select2 and its dependent scripts are only
registered on the specific pages where they need to be enqueued.

Fixes #48.